Adding integration test orchestrator
Problem: we needed to be able to pass a custom initramfs to a test instead of having the test use the generic initramfs.

I couldn’t find a better way of passing the initramfs to a test besides passing it as a flag. But that gets annoying if you  want to run all tests at once, so I wanted to have a test orchestrator that made it easy to run all the tests at once, which also let you specify specific initramfs.

So the way this would work is that main_test.go has a single ‘parent’ test, with all the integration tests running as subtests (this gives us the naming hierarchy so that individual tests can be run as `go test -test.run=TestIntegration/Foo`).  This then allows us to pass in our own initramfs’ (eg. `go test -TestDhclient_Initramfs=PATH -TestScript_Initramfs=PATH`)

Things I don’t like about this: it splits the test into 2 places, an actual test and the entry in the main_test.go table. The env variable feels kind of messy. I wanted to just be able to split out the entire initramfs building into a separate piece but that requires splitting out the test’s u-root build opts again.

If there’s a better way this could be done that would be great!

Signed-off-by: pallaud <plaud@google.com>